crossplatformability improvements

+++ source/client/cl_parse.c
@@ -627,7 +627,7 @@ static void CL_InitServerDownload( const char *filename, int size, unsigned chec
 	cls.download.timeout = Sys_Milliseconds() + 3000;
 	cls.download.retries = 0;
 
-	CL_AddReliableCommand( va( "nextdl \"%s\" %i", cls.download.name, cls.download.offset ) );
+	CL_AddReliableCommand( va( "nextdl \"%s\" %zu", cls.download.name, cls.download.offset ) );
 }
 
 /*
@@ -710,7 +710,7 @@ static void CL_RetryDownload( void )
 	else
 	{
 		cls.download.timeout = Sys_Milliseconds() + 3000;
-		CL_AddReliableCommand( va( "nextdl \"%s\" %i", cls.download.name, cls.download.offset ) );
+		CL_AddReliableCommand( va( "nextdl \"%s\" %zu", cls.download.name, cls.download.offset ) );
 	}
 }
 
@@ -854,7 +854,7 @@ static void CL_ParseDownload( msg_t *msg )
 		cls.download.timeout = Sys_Milliseconds() + 3000;
 		cls.download.retries = 0;
 
-		CL_AddReliableCommand( va( "nextdl \"%s\" %i", cls.download.name, cls.download.offset ) );
+		CL_AddReliableCommand( va( "nextdl \"%s\" %zu", cls.download.name, cls.download.offset ) );
 	}
 	else
 	{
@@ -884,7 +884,7 @@ static void CL_ParseServerData( msg_t *msg )
 {
 	const char *str, *gamedir;
 	int i, sv_bitflags, numpure;
-	int http_portnum;
+	unsigned short http_portnum;
 	bool old_sv_pure;
 
 	Com_DPrintf( "Serverdata packet received.\n" );
@@ -1311,9 +1311,9 @@ void CL_ParseServerMessage( msg_t *msg )
 		if( cl_debug_serverCmd->integer & 4 )
 		{
 			if( cmd == -1 )
-				Com_Printf( "%3i:CMD %i %s\n", msg->readcount-1, cmd, "EOF" );
+				Com_Printf( "%3zu:CMD %i %s\n", msg->readcount-1, cmd, "EOF" );
 			else
-				Com_Printf( "%3i:CMD %i %s\n", msg->readcount-1, cmd, !svc_strings[cmd] ? "bad" : svc_strings[cmd] );
+				Com_Printf( "%3zu:CMD %i %s\n", msg->readcount-1, cmd, !svc_strings[cmd] ? "bad" : svc_strings[cmd] );
 		}
 
 		if( cmd == -1 )
